aws/pdw_downloader/requirements.txt
snowflake-connector-python==2.9.0
pyarrow==8.0.0
pandas==1.5.2